backend/pyproject.toml
[project]
name = "rand-lottery-api"
version = "0.1.0"
description = "FastAPI backend for Rand Lottery"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.5.2",
  "SQLAlchemy>=2.0.36",
  "asyncpg>=0.30.0",
  "alembic>=1.13.3",
  "psycopg[binary]>=3.2.3",
  "python-multipart>=0.0.9",
  "httpx>=0.27.2",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "requests>=2.31.0",
  "requests_oauthlib>=1.3.1",
  "aiosmtplib>=2.0.2",
  "google-auth>=2.35.0",
  "email-validator>=2.2.0",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
